+++ encrypt.cu
@@ -152,7 +152,7 @@ __device__ bool is_valid_key(u8 *key, u16 key_length, u8 *salt, u16 salt_length,
     return true;
 }
 
-__device__ u8 add(int64_t test_pwd_num[MAX_PWD_LENGTH], u64 a, u64 legal_chars_length)
+__device__ u64 add(int64_t test_pwd_num[MAX_PWD_LENGTH], u64 a, u64 legal_chars_length)
 {
     // add test_pwd_num by a
     u64 carry = 0;
